Core: switched tabs + added quickbar for options (#145)
* core: switched tabs and gave them static posValues. also added QuicActionBar in tab "options"

* core: made QuickActionsBar a generic jsx component to be used in any sidebar.

* options sidebar now also has the quick actions.

---------
